# Optimize State merge 